Refactor Vue Croppie
